backend/requirements/dev.txt
-r base.txt

# Testing
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.1
faker==26.0.0

# Linting / Formatting
ruff==0.5.5
mypy==1.11.1
django-stubs==5.0.4
djangorestframework-stubs==3.15.0

# Debug
ipython==8.26.0
django-debug-toolbar==4.4.6
